refactor: split Spark type parsing out of the shared sql package
